:sparkles: feature setting initialization function (#7)
