+++ package.json
@@ -18,10 +18,8 @@
     "react-datepicker": "^4.10.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
-    "react-infinite-scroll-component": "^6.1.0",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.1",
-    "reactive-button": "^1.3.13",
     "styled-components": "^5.3.6",
     "styled-reset": "^4.4.5"
   },
